LOGIN WERKT

+++ webapp/webapp/Pages/Menu/SMenu.cshtml.cs
@@ -11,16 +11,26 @@ namespace webapp.Pages.Menu
     public class SMenuModel : PageModel
     {
         public string postit { get; set; }
+        public int LU { get; set; }
         public static void OnGet()
         {
 
         }
         public void OnPost()
         {
+            string loguit = Request.Form[nameof(LU)];
+            if(loguit != null)
+            {
+                Sessie.GetInstance.setLoginUserID(-1);
+                Response.Redirect("../Index");
+            }
+            else
+            {
+                string postit2 = Request.Form[nameof(postit)];
+                int id = Convert.ToInt32(postit2);
+                new DBConnection().Send("DELETE FROM `projectcdb`.`vraag` WHERE(`vraagID` = '" + id + "');");
+            }
 
-            string postit2 = Request.Form[nameof(postit)];
-            int id = Convert.ToInt32(postit2);
-            new DBConnection().Send("DELETE FROM `projectcdb`.`vraag` WHERE(`vraagID` = '" + id + "');");
         }
     }
 }
